+++ xarray/plot/facetgrid.py
@@ -267,9 +267,11 @@ class FacetGrid:
                 mappable = func(subset, x=x, y=y, ax=ax, **func_kwargs)
                 self._mappables.append(mappable)
 
+                animate = kwargs.pop('animate', None)
         _, _, hueplt, xlabel, ylabel, huelabel = _infer_line_data(
             darray=self.data.loc[self.name_dicts.flat[0]],
-            x=x, y=y, hue=func_kwargs['hue'])
+            x=x, y=y, hue=func_kwargs['hue'], animate=animate,
+            linestyle=func_kwargs.get('linestyle', ''))
 
         self._hue_var = hueplt
         self._hue_label = huelabel